python/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
asgiref==3.7.2
asn1crypto==1.4.0
awscli==1.27.150
bcrypt==4.0.1
blinker==1.7.0
botocore==1.29.150
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contextvars==2.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
dnspython==2.4.2
docutils==0.16
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.103.2
fasteners==0.14.1
Flask==3.0.2
Flask-Cors==4.0.0
future==0.18.2
greenlet==3.0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
httptools==0.6.0
idna==3.4
immutables==0.20
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.5
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pupil-labs-uvc==1.0.0rc1
pyasn1==0.5.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.7.2
s3transfer==0.6.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
starlette==0.27.0
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.18.0
wadllib==1.3.6
watchfiles==0.21.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
